feature : jwt

+++ greenlight-core/src/main/java/com/greenlight/global/presentation/response/SignUpResponse.java
@@ -1,20 +1,21 @@
 package com.greenlight.global.presentation.response;
 
-import com.greenlight.global.application.processor.MemberSignUpProcessor;
 import lombok.Getter;
 import lombok.ToString;
+
+import com.greenlight.global.application.processor.MemberSignUpProcessor;
 
 @Getter
 @ToString
 public class SignUpResponse {
 
-	private Long memberId;
+	private int memberId;
 
 	public static SignUpResponse from(MemberSignUpProcessor.Result result) {
 		return new SignUpResponse(result.getMemberId());
 	}
 
-	public SignUpResponse(Long memberId) {
+	public SignUpResponse(int memberId) {
 		this.memberId = memberId;
 	}
 }
